Create RBAC to grant the OLM operator read-only access to Secrets in `kube-system`
When enabling the BGP control plane feature, the OLM operator needs to
create a Role and RoleBinding for read-only access to Secrets in the
`kube-system` namespace. In Cilium <= 1.14, there's no way to change the
namespace which the OLM operator configures as the source for BGP
peering secrets to something other than `kube-system`.

However, if Cilium is installed in a namespace other than `kube-system`,
such as `cilium`, the OLM operator doesn't have sufficient permissions
to create the required Role and RoleBinding in the `kube-system`
namespace.

This commit generates RBAC rules which ensure that the OLM operator can
create the required RBAC rules in the kube-system namespace.