Patch OLM RBAC to allow deploying Hubble certgen cronjob
The Cilium Helm chart for cilium <= 1.15 uses a ClusterRole and
ClusterRoleBinding for the Hubble certgen CronJob. This is changed to a
Role and RoleBinding for the Cilium 1.16 Helm chart.

The OLM operator doesn't have permissions to create the ClusterRole and
ClusterRoleBinding out of the box, so we patch the OLM operator
ClusterRole to contain the rules that need to be created for the certgen
CronJob.