Add support to configure a custom node selector for the shadow ranges DaemonSet
This can be useful for example when using the component on an OpenShift
4 cluster where the machine config pool of the egress nodes contains
additional nodes which aren't assigned a shadow range.

In that case, we must be able to ensure that the DaemonSet runs on all
nodes so that script can access the config map from all nodes, while
ensuring that the generated egress gateway policies don't select any
nodes that don't have a shadow range assigned.