Extend `NamespaceEgressPolicy` to support BGP-backed egress IPs
For BGP-backed egress IPs, we don't configure the policy's egress IP
based on the computed interface, but instead directly specify the egress
IP in the policy. Additionally, we always set `maxGatewayNodes: 1` for
BGP-backed egress IPs.

Note that it's still the responsibility of the cluster operator to
ensure that all possible BGP-backed egress IPs are assigned to an
interface on each egress node.