Cilium 1.17 doesn't like having duplicated hubble metrics configs
This commit adds a bit of logic which takes the last entry in
`hubble.metrics.enabled` for each metrics type (i.e. making the part
before the first `|` unique in the entries).

By taking the last entry for each type we allow users to customize
metrics which are enabled by default through the config hierarchy by
simply providing an additional entry for the same metrics type.