Add ArgoCD pre-sync job for migrating from cilium-ee-olm to clife
The sync hook executes the steps which are documented on
docs.isovalent.com for migrating from cilium-ee-olm to clife.

After deleting the `ciliumconfigs.cilium.io` CRD, we don't recreate the
CRD and config from the pre-sync job, since that's not enough for ArgoCD
to then be able to apply the new ciliumconfig resource.

As far as we've been able to determine the only way to avoid the ArgoCD
sync getting stuck is to delete the
`syn-argocd-application-controller-0` pod to clear any conflicting cache
entries for `ciliumconfigs.cilium.io`.