Distinguish between manifests version and image version when rendering config
We want to decide the correct CRD API version based on the image
version, which can be overridden in the Helm values. However, we still
want to make decisions for rendering the OLM manifests based on the
manifests version.

To achieve this, we rename the previous `util.version` to
`util.manifestsVersion` and introduce a new `util.version` which infers
the version from Helm value `image.tag` if it's set and falls back to
`util.manifestsVersion` otherwise.

The rest of the component is updated to use `manifestsVersion` for
version checks which are only relevant for rendering the OLM manifests
or Helm chart.